+++ lib/features/history/widgets/detection_actions.dart
@@ -300,4 +300,3 @@ class _OverflowRow extends StatelessWidget {
     );
   }
 }
-